+++ src/Graphics/GL/Terrain.cpp
@@ -51,35 +51,35 @@ GL::Terrain::Terrain(std::uint32_t width, std::uint32_t height) : ITerrain(width
 
 	m_vbo = std::make_unique<GL::Buffer>(ARRAY_BUFFER);
 	m_vbo->Bind();
-	m_vbo->Update(m_vertices.size() * sizeof(glm::vec3), &m_vertices[0]);
+	m_vbo->Update((unsigned int)m_vertices.size() * sizeof(glm::vec3), &m_vertices[0]);
 
 	glEnableVertexAttribArray(0);
 	glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 0, (void*)0);
 
 	m_uvbo = std::make_unique<GL::Buffer>(ARRAY_BUFFER);
 	m_uvbo->Bind();
-	m_uvbo->Update(m_uvs.size() * sizeof(glm::vec2), &m_uvs[0]);
+	m_uvbo->Update((unsigned int)m_uvs.size() * sizeof(glm::vec2), &m_uvs[0]);
 	
 	glEnableVertexAttribArray(1);
 	glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, 0, (void*)0);
 
 	m_nbo = std::make_unique<GL::Buffer>(ARRAY_BUFFER);
 	m_nbo->Bind();
-	m_nbo->Update(m_normals.size() * sizeof(glm::vec3), &m_normals[0]);
+	m_nbo->Update((unsigned int)m_normals.size() * sizeof(glm::vec3), &m_normals[0]);
 
 	glEnableVertexAttribArray(2);
 	glVertexAttribPointer(2, 3, GL_FLOAT, GL_FALSE, 0, (void*)0);
 
 	m_mbo = std::make_unique<GL::Buffer>(ARRAY_BUFFER);
 	m_mbo->Bind();
-	m_mbo->Update(m_materials.size() * sizeof(std::uint32_t), &m_materials[0]);
+	m_mbo->Update((unsigned int)m_materials.size() * sizeof(std::uint32_t), &m_materials[0]);
 
 	glEnableVertexAttribArray(3);
 	glVertexAttribPointer(3, 3, GL_FLOAT, GL_FALSE, 0, (void*)0);
 
 	m_fbo = std::make_unique<GL::Buffer>(ELEMENT_ARRAY_BUFFER);
 	m_fbo->Bind();
-	m_fbo->Update(m_faces.size() * sizeof(std::uint32_t), &m_faces[0]);
+	m_fbo->Update((unsigned int)m_faces.size() * sizeof(std::uint32_t), &m_faces[0]);
 
 }
